refactor(renderer):  index export for backdrop
