+++ src/renderer/src/components/Drawer.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, useEffect, useRef } from "react";
 import { createPortal } from "react-dom";
 import CloseIcon from "~/assets/icons/xmark.svg?react";
-import { useAttachBackdrop } from "~/features/Backdrop/useAttachBackdrop";
+import { useAttachBackdrop } from "~/features/Backdrop";
 import { usePortalRoot } from "~/hooks/dom/usePortalRoot";
 import { classed } from "~/lib/classed";
 import { Button } from "./Button";
