+++ src/renderer/src/components/Modal.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useCallback } from "react";
 import { createPortal } from "react-dom";
-import { useAttachBackdrop } from "~/features/Backdrop/useAttachBackdrop";
+import { useAttachBackdrop } from "~/features/Backdrop";
 import { usePortalRoot } from "~/hooks/dom/usePortalRoot";
 import { classed } from "~/lib/classed";
 import { Button } from "./Button";
